extracted channel UI components into separate script class
