+++ unity/Assets/Ably/Examples/AblyChannelUiConsole.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a6e2024162482c84a96750f78d375aa5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
